+++ CHANGELOG.md
@@ -15,4 +15,5 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
     - Long-story-short, aquatic worldgen features can be toggled in-game without needing to rejoin the world!
   - Please monitor worldgen performance and tell us if you notice any issues. This is an incredibly important feature to us.
     - If no performance issues are found, we will slowly expand this new system to other worldgen features.
+- All of Wilder Wild's block sound type overwrites are now data-driven, per FrozenLib 2.5's changes.
 - Slightly decreased the average size of Mesoglea Caves.